requirements.txt
praw==6.5.1
google-cloud-bigquery==1.24.0